+++ firmware/src/BusinessLogic/nodeControl.h
@@ -14,74 +14,10 @@ Copyright (c) 2019 Panda Team
 */
 
 #include <memory>
-#include "ADmux.h"
+#include "BaseMesChannel.h"
 #include "zerocal_man.h"
 #include "json_evsys.h"
 #include "RawBinStorage.h"
-#include "DAC.h"
-#include "Pin.h"
-#include "DataVis.h"
-
-class nodeControl;
-class CMesChannel
-{
-friend class nodeControl;
-public:
-
-    enum mes_mode{
-
-        Voltage=0,
-        Current
-    };
-
-    //interface:
-    inline int GetADCmesRawVal(){
-
-        return m_pADC->GetRawBinVal();
-    }
-
-    virtual void Enable(bool bHow){
-
-        m_bEnabled=bHow;
-    }
-    virtual void SetMesMode(mes_mode nMode){
-
-        m_MesMode=nMode;
-    }
-    virtual void SetAmpGain(float GainValue){
-
-        m_ActualAmpGain=GainValue;
-    }
-    inline float GetActualAmpGain()
-    {
-        return m_ActualAmpGain;
-    }
-
-protected:
-    nodeControl *m_pCont=nullptr;
-
-    bool m_bEnabled=false;
-    mes_mode m_MesMode=mes_mode::Voltage;
-    float m_ActualAmpGain=1.0f;
-
-    std::shared_ptr<CAdc> m_pADC;
-    std::shared_ptr<CDac> m_pDAC;
-    CDataVis m_VisChan;
-
-    void Update(){
-
-        m_VisChan.Update( m_pADC->GetRawBinVal() );
-    }
-
-public:
-    CMesChannel(const std::shared_ptr<CAdc> &pADC,  const std::shared_ptr<CDac> &pDAC,  CView::vischan nCh) : m_VisChan(nCh){
-
-        m_pADC=pADC;
-        m_pDAC=pDAC;
-    }
-
-
-};
 
 
 /*!
@@ -117,6 +53,9 @@ protected:
         CRawBinStorage m_PersistStorage;
 
         int m_GainSetting=0;
+
+        bool m_BridgeSetting=false;
+        int  m_SecondarySetting=0;
 
         /*!
          * \brief Holds Current Setting (mockup)
